[proliferate] consolidate subagents mcp server
