+++ anyharness/crates/anyharness-lib/src/api/http/subagents.rs
@@ -13,7 +13,7 @@ use serde_json::Value;
 use super::access::assert_workspace_mutable;
 use super::error::ApiError;
 use crate::app::AppState;
-use crate::sessions::subagents::mcp::handle_json_rpc;
+use crate::sessions::subagents::mcp_server::server::handle_json_rpc;
 use crate::sessions::subagents::service::SubagentError;
 use crate::workspaces::operation_gate::WorkspaceOperationKind;
