+++ anyharness/crates/anyharness-lib/src/sessions/subagents/mcp_server/auth.rs
