+++ anyharness/crates/anyharness-lib/src/sessions/subagents/mod.rs
@@ -1,7 +1,5 @@
 pub mod hooks;
-pub mod mcp;
-pub mod mcp_auth;
-mod mcp_protocol;
+pub mod mcp_server;
 pub mod model;
 pub mod service;
 pub mod store;
